Update README
- Updates TemporalGPs-related note to be more accurate.
- Adds Stheno.jl reference now that it's inside the org.
- Fix capitalisation

I think we should introduce an informal rule that packages which implement types that implement the APIs specified in this package, or which implements functionality built directly on top of APIs in this package (eg. something for handling non-Gaussian likelihoods that utilises AbstractGPs and / or LatentGPs), should be included in this section.